+++ app/widgets/Avatar/Avatar.php
@@ -13,6 +13,7 @@ class Avatar extends \Movim\Widget\Base
 
         $this->registerEvent('avatar_get_handle', 'onGetAvatar');
         $this->registerEvent('avatar_set_handle', 'onSetAvatar');
+        $this->registerEvent('avatar_set_pubsub', 'onSetBanner');
         $this->registerEvent('avatar_set_errorfeaturenotimplemented', 'onMyAvatarError');
         $this->registerEvent('avatar_set_errorbadrequest', 'onMyAvatarError');
         $this->registerEvent('avatar_set_errornotallowed', 'onMyAvatarError');
@@ -21,6 +22,16 @@ class Avatar extends \Movim\Widget\Base
     public function onSetAvatar($packet)
     {
         $this->ajaxGetAvatar();
+        $this->rpc('MovimTpl.fill', '#avatar', $this->prepareForm());
+    }
+
+    public function onSetBanner($packet)
+    {
+        global $loop;
+
+        $loop->addTimer(2, function () {
+            $this->rpc('MovimTpl.fill', '#avatar', $this->prepareForm());
+        });
     }
 
     public function onGetAvatar($packet)
@@ -48,6 +59,13 @@ class Avatar extends \Movim\Widget\Base
         $view = $this->tpl();
         $view->assign('me', \App\Contact::firstOrNew(['id' => $this->user->id]));
         Dialog::fill($view->draw('_avatar_form'));
+    }
+
+    public function ajaxGetBannerForm()
+    {
+        $view = $this->tpl();
+        $view->assign('me', \App\Contact::firstOrNew(['id' => $this->user->id]));
+        Dialog::fill($view->draw('_avatar_banner_form'));
     }
 
     public function ajaxGetAvatar()
@@ -79,5 +97,28 @@ class Avatar extends \Movim\Widget\Base
         $p->remove();
 
         $p->remove('jpeg');
+    }
+
+    public function ajaxBannerSubmit($banner)
+    {
+        if (empty($banner->photobin->value)) return;
+
+        $key = $this->user->id.'banner';
+
+        $p = new Image;
+        $p->fromBase($banner->photobin->value);
+        $p->setKey($key);
+        $p->save(false, false, 'jpeg', 60);
+
+        // Reload
+        $p->load('jpeg');
+
+        $r = new Set;
+        $r->setNode('urn:xmpp:movim-banner:0')
+          ->setUrl(Image::getOrCreate($key, false, false, 'jpeg', true))
+          ->setWidthMetadata(1280)
+          ->setHeightMetadata(320)
+          ->setData($p->toBase())
+          ->request();
     }
 }