+++ lib/moxl/src/Xec/Action/Pubsub/GetItems.php
@@ -81,12 +81,14 @@ class GetItems extends Errors
 
                 if ($i && $i->avatarhash !== (string)$item->metadata->info->attributes()->id) {
                     $p = new Image;
-                    $p->fromURL((string)$item->metadata->info->attributes()->url);
-                    $p->setKey((string)$item->metadata->info->attributes()->id);
-                    $p->save();
 
-                    $i->avatarhash = (string)$item->metadata->info->attributes()->id;
-                    $i->save();
+                    if ($p->fromURL((string)$item->metadata->info->attributes()->url)) {
+                        $p->setKey((string)$item->metadata->info->attributes()->id);
+                        $p->save();
+
+                        $i->avatarhash = (string)$item->metadata->info->attributes()->id;
+                        $i->save();
+                    }
                 }
             }
         }