+++ public/scripts/movim_avatar.js
@@ -1,5 +1,5 @@
 var MovimAvatar = {
-    file : function(files, formname) {
+    file : function(files, formname, width = 350, height = 350) {
         var f = files[0];
         if (!f.type.match(/image.*/)) {
           console.log("Not a picture !");
@@ -10,7 +10,7 @@ var MovimAvatar = {
 
             reader.onload = function (ev) {
                 MovimUtils.getOrientation(f, function(orientation) {
-                    MovimAvatar.preview(ev.target.result, orientation, formname);
+                    MovimAvatar.preview(ev.target.result, orientation, formname, width, height);
                 });
             };
         };
@@ -19,9 +19,10 @@ var MovimAvatar = {
         document.querySelector('form[name=' + formname + '] img').src = '';
         document.querySelector('form[name=' + formname + '] input[name="photobin"]').value = '';
     },
-    preview : function(src, orientation, formname) {
+    preview : function(src, orientation, formname, setWidth, setHeight) {
         var canvas = document.createElement('canvas');
-        width = height = canvas.width = canvas.height = 350;
+        width = canvas.width = setWidth;
+        height = canvas.height = setHeight;
 
         var image = new Image();
         image.src = src;
@@ -39,16 +40,7 @@ var MovimAvatar = {
                 default: ctx.transform(1, 0, 0, 1, 0, 0);
             }
 
-            if (image.width == image.height) {
-                ctx.drawImage(image, 0, 0, width, height);
-            } else {
-                minVal = parseInt(Math.min(image.width, image.height));
-                if (image.width > image.height) {
-                    ctx.drawImage(image, (parseInt(image.width) - minVal) / 2, 0, minVal, minVal, 0, 0, width, height);
-                } else {
-                    ctx.drawImage(image, 0, (parseInt(image.height) - minVal) / 2, minVal, minVal, 0, 0, width, height);
-                }
-            }
+            MovimUtils.drawImageProp(ctx, image);
 
             var base64 = canvas.toDataURL('image/jpeg', 0.95);